[fix] move securityConfig, JWT dir to global
