+++ src/main/java/com/be/rebook/domain/members/service/ReissueService.java
@@ -2,9 +2,10 @@ package com.be.rebook.domain.members.service;
 
 import com.be.rebook.domain.members.repository.RefreshTokensRepository;
 import com.be.rebook.domain.members.entity.RefreshTokens;
-import com.be.rebook.domain.members.jwt.JWTUtil;
 import com.be.rebook.global.exception.BaseException;
 import com.be.rebook.global.exception.ErrorCode;
+import com.be.rebook.global.security.jwt.JWTUtil;
+
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
