+++ src/main/java/com/be/rebook/global/security/jwt/LoginFilter.java
@@ -1,4 +1,4 @@
-package com.be.rebook.domain.members.jwt;
+package com.be.rebook.global.security.jwt;
 
 import com.be.rebook.domain.members.entity.RefreshTokens;
 import com.be.rebook.domain.members.repository.RefreshTokensRepository;
